requirements.txt
streamlit
deepface
opencv-python
numpy
Pillow
tensorflow
